+++ integration_test/cases/type.exs
@@ -1,7 +1,7 @@
 defmodule Ecto.Integration.TypeTest do
   use Ecto.Integration.Case, async: Application.compile_env(:ecto, :async_integration_tests, true)
 
-  alias Ecto.Integration.{Custom, Item, ItemColor, Order, Post, User, Tag, Usec}
+  alias Ecto.Integration.{Comment, Custom, Item, ItemColor, Order, Post, User, Tag, Usec}
   alias Ecto.Integration.TestRepo
   import Ecto.Query
 
@@ -114,7 +114,8 @@ defmodule Ecto.Integration.TypeTest do
   end
 
   test "tagged types" do
-    TestRepo.insert!(%Post{visits: 12})
+    %{id: post_id} = TestRepo.insert!(%Post{visits: 12})
+    TestRepo.insert!(%Comment{text: "#{post_id}", post_id: post_id})
 
     # Numbers
     assert [1]   = TestRepo.all(from Post, select: type(^"1", :integer))
@@ -141,6 +142,11 @@ defmodule Ecto.Integration.TypeTest do
     # Comparison expression
     assert [12] = TestRepo.all(from p in Post, select: type(coalesce(p.visits, 0), :integer))
     assert [1.0] = TestRepo.all(from p in Post, select: type(coalesce(p.intensity, 1.0), :float))
+
+    # parent_as/1
+    child = from c in Comment, where: type(parent_as(:posts).id, :string) == c.text, select: c.post_id
+    query = from p in Post, as: :posts, where: p.id in subquery(child), select: p.id
+    assert [post_id] == TestRepo.all(query)
   end
 
   test "binary id type" do
@@ -294,7 +300,7 @@ defmodule Ecto.Integration.TypeTest do
   @tag :map_type
   @tag :json_extract_path
   test "json_extract_path with primitive values" do
-    order = %Order{meta:
+    order = %Order{metadata:
       %{
         :id => 123,
         :time => ~T[09:00:00],
@@ -308,44 +314,44 @@ defmodule Ecto.Integration.TypeTest do
 
     order = TestRepo.insert!(order)
 
-    assert TestRepo.one(from o in Order, select: o.meta["id"]) == 123
-    assert TestRepo.one(from o in Order, select: o.meta["bad"]) == nil
-    assert TestRepo.one(from o in Order, select: o.meta["bad"]["bad"]) == nil
+    assert TestRepo.one(from o in Order, select: o.metadata["id"]) == 123
+    assert TestRepo.one(from o in Order, select: o.metadata["bad"]) == nil
+    assert TestRepo.one(from o in Order, select: o.metadata["bad"]["bad"]) == nil
 
     field = "id"
-    assert TestRepo.one(from o in Order, select: o.meta[^field]) == 123
-    assert TestRepo.one(from o in Order, select: o.meta["time"]) == "09:00:00"
-    assert TestRepo.one(from o in Order, select: o.meta["'single quoted'"]) == "bar"
-    assert TestRepo.one(from o in Order, select: o.meta["';"]) == nil
-    assert TestRepo.one(from o in Order, select: o.meta["\"double quoted\""]) == "baz"
-    assert TestRepo.one(from o in Order, select: o.meta["enabled"]) == true
-    assert TestRepo.one(from o in Order, select: o.meta["extra"][0]["enabled"]) == false
+    assert TestRepo.one(from o in Order, select: o.metadata[^field]) == 123
+    assert TestRepo.one(from o in Order, select: o.metadata["time"]) == "09:00:00"
+    assert TestRepo.one(from o in Order, select: o.metadata["'single quoted'"]) == "bar"
+    assert TestRepo.one(from o in Order, select: o.metadata["';"]) == nil
+    assert TestRepo.one(from o in Order, select: o.metadata["\"double quoted\""]) == "baz"
+    assert TestRepo.one(from o in Order, select: o.metadata["enabled"]) == true
+    assert TestRepo.one(from o in Order, select: o.metadata["extra"][0]["enabled"]) == false
 
     # where
-    assert TestRepo.one(from o in Order, where: o.meta["id"] == 123, select: o.id) == order.id
-    assert TestRepo.one(from o in Order, where: o.meta["id"] == 456, select: o.id) == nil
-    assert TestRepo.one(from o in Order, where: o.meta["code"] == "good", select: o.id) == order.id
-    assert TestRepo.one(from o in Order, where: o.meta["code"] == "bad", select: o.id) == nil
-    assert TestRepo.one(from o in Order, where: o.meta["enabled"] == true, select: o.id) == order.id
-    assert TestRepo.one(from o in Order, where: o.meta["extra"][0]["enabled"] == false, select: o.id) == order.id
+    assert TestRepo.one(from o in Order, where: o.metadata["id"] == 123, select: o.id) == order.id
+    assert TestRepo.one(from o in Order, where: o.metadata["id"] == 456, select: o.id) == nil
+    assert TestRepo.one(from o in Order, where: o.metadata["code"] == "good", select: o.id) == order.id
+    assert TestRepo.one(from o in Order, where: o.metadata["code"] == "bad", select: o.id) == nil
+    assert TestRepo.one(from o in Order, where: o.metadata["enabled"] == true, select: o.id) == order.id
+    assert TestRepo.one(from o in Order, where: o.metadata["extra"][0]["enabled"] == false, select: o.id) == order.id
   end
 
   @tag :map_type
   @tag :json_extract_path
   test "json_extract_path with arrays and objects" do
-    order = %Order{meta: %{tags: [%{name: "red"}, %{name: "green"}]}}
+    order = %Order{metadata: %{tags: [%{name: "red"}, %{name: "green"}]}}
     order = TestRepo.insert!(order)
 
-    assert TestRepo.one(from o in Order, select: o.meta["tags"][0]["name"]) == "red"
-    assert TestRepo.one(from o in Order, select: o.meta["tags"][99]["name"]) == nil
+    assert TestRepo.one(from o in Order, select: o.metadata["tags"][0]["name"]) == "red"
+    assert TestRepo.one(from o in Order, select: o.metadata["tags"][99]["name"]) == nil
 
     index = 1
-    assert TestRepo.one(from o in Order, select: o.meta["tags"][^index]["name"]) == "green"
+    assert TestRepo.one(from o in Order, select: o.metadata["tags"][^index]["name"]) == "green"
 
     # where
-    assert TestRepo.one(from o in Order, where: o.meta["tags"][0]["name"] == "red", select: o.id) == order.id
-    assert TestRepo.one(from o in Order, where: o.meta["tags"][0]["name"] == "blue", select: o.id) == nil
-    assert TestRepo.one(from o in Order, where: o.meta["tags"][99]["name"] == "red", select: o.id) == nil
+    assert TestRepo.one(from o in Order, where: o.metadata["tags"][0]["name"] == "red", select: o.id) == order.id
+    assert TestRepo.one(from o in Order, where: o.metadata["tags"][0]["name"] == "blue", select: o.id) == nil
+    assert TestRepo.one(from o in Order, where: o.metadata["tags"][99]["name"] == "red", select: o.id) == nil
   end
 
   @tag :map_type
@@ -355,6 +361,14 @@ defmodule Ecto.Integration.TypeTest do
     TestRepo.insert!(order)
 
     assert TestRepo.one(from o in Order, select: o.items[0]["valid_at"]) == "2020-01-01"
+  end
+
+  @tag :map_type
+  @tag :json_extract_path
+  test "json_extract_path with custom field source" do
+    order = TestRepo.insert!(%Order{metadata: %{tags: [%{name: "red"}, %{name: "green"}]}})
+
+    assert TestRepo.one(from o in Order, where: o.metadata["tags"][0]["name"] == "red", select: o.id) == order.id
   end
 
   @tag :map_type