+++ lib/ecto/repo.ex
@@ -1158,6 +1158,9 @@ defmodule Ecto.Repo do
       from(p in Post, where: p.id < 10, update: [set: [title: fragment("upper(?)", ^new_title)]])
       |> MyRepo.update_all([])
 
+      from(p in Post, where: p.id < 10, update: [set: [visits: p.visits * 1000]])
+      |> MyRepo.update_all([])
+
   """
   @doc group: "Query API"
   @callback update_all(