+++ test/ecto/query/builder/select_test.exs
@@ -9,11 +9,23 @@ defmodule Ecto.Query.Builder.SelectTest do
     defstruct [:title]
   end
 
+  defmodule Comment do
+    use Ecto.Schema
+
+    @primary_key false
+    schema "comments" do
+      field :title, :string
+      field :likes, :integer
+      field :dislikes, :integer, load_in_query: false
+    end
+  end
+
   defp params_acc(opts \\ []) do
     params = opts[:params] || []
     take = opts[:take] || %{}
     subqueries = opts[:subqueries] || []
-    {params, %{take: take, subqueries: subqueries}}
+    aliases = opts[:aliases] || %{}
+    {params, %{take: take, subqueries: subqueries, aliases: aliases}}
   end
 
   describe "escape" do
@@ -84,6 +96,71 @@ defmodule Ecto.Query.Builder.SelectTest do
     test "raises on mixed fields and interpolation" do
       assert_raise Ecto.Query.CompileError, ~r"Cannot mix fields with interpolations", fn ->
         escape(quote do [:foo, ^:bar] end, [], __ENV__)
+      end
+    end
+
+    test "supports aliasing a selected value with selected_as/2" do
+      escaped_alias = {:selected_as, [], [{{:., [], [{:&, [], [0]}, :id]}, [], []}, :ident]}
+
+      # single field
+      query = from p in "posts", select: selected_as(p.id, :ident)
+      assert escaped_alias == query.select.expr
+
+      query = select("posts", [p], selected_as(p.id, :ident))
+      assert escaped_alias == query.select.expr
+
+      # maps
+      query = from p in "posts", select: %{id: selected_as(p.id, :ident)}
+      assert {:%{}, [], [id: escaped_alias]} == query.select.expr
+
+      query = select("posts", [p], %{id: selected_as(p.id, :ident)})
+      assert {:%{}, [], [id: escaped_alias]} == query.select.expr
+
+      # structs
+      query = from p in "posts", select: %{p | id: selected_as(p.id, :ident)}
+      assert {:%{}, [], [{:|, [], [{:&, [], [0]}, [id: escaped_alias]]}]} == query.select.expr
+
+      query = select("posts", [p], %{p | id: selected_as(p.id, :ident)})
+      assert {:%{}, [], [{:|, [], [{:&, [], [0]}, [id: escaped_alias]]}]} == query.select.expr
+
+      # keyword lists
+      query = from p in "posts", select: [id: selected_as(p.id, :ident)]
+      assert [{:{}, [], [:id, escaped_alias]}] == query.select.expr
+
+      query = select("posts", [p], [id: selected_as(p.id, :ident)])
+      assert [{:{}, [], [:id, escaped_alias]}] == query.select.expr
+    end
+
+    test "supports aliasing a selected value in select_merge with selected_as/2" do
+      escaped_alias = {:selected_as, [], [{{:., [], [{:&, [], [0]}, :id]}, [], []}, :ident]}
+
+      query = from p in "posts", select: p.visits, select_merge: %{id: selected_as(p.id, :ident)}
+      assert {:merge, [], [{{:., [], [{:&, [], [0]}, :visits]}, [], []}, {:%{}, [], [id: escaped_alias]}]} == query.select.expr
+    end
+
+    test "raises if name given to selected_as/2 is not an atom" do
+      message = "selected_as/2 expects `name` to be an atom, got `\"ident\"`"
+
+      assert_raise Ecto.Query.CompileError, message, fn ->
+        escape(quote do selected_as(p.id, "ident") end, [], __ENV__)
+      end
+    end
+
+    test "raises if the name given to selected_as/2 already exists" do
+      message = "the alias `:ident` has been specified more than once using `selected_as/2`"
+
+      assert_raise Ecto.Query.CompileError, message, fn ->
+        select_expr = quote do %{id: selected_as(p.id, :ident), id2: selected_as(p.id, :ident)} end
+        escape(select_expr, [p: 0], __ENV__)
+      end
+    end
+
+    test "raises if selected_as/2 is not at the root of the select statement" do
+      message = ~r/selected_as\/2 can only be used at the root of a select statement/
+
+      assert_raise Ecto.Query.CompileError, message, fn ->
+        select_expr = quote do coalesce(selected_as(p.visits, :v), 0) end
+        escape(select_expr, [p: 0], __ENV__)
       end
     end
   end
@@ -435,19 +512,33 @@ defmodule Ecto.Query.Builder.SelectTest do
     end
 
     test "with take" do
-      # On select
-      query = from p in "posts", select: p, select_merge: [:title]
+      # On select with schemaless source
+      query = from c in "comments", select: c, select_merge: [:title]
 
       assert Macro.to_string(query.select.expr) == "&0"
       assert query.select.params == []
       assert query.select.take == %{}
 
-      # On take
-      query = from p in "posts", select: [:body], select_merge: [:title]
+      # On select with schema
+      query = from c in Comment, select: c, select_merge: [:dislikes]
 
       assert Macro.to_string(query.select.expr) == "&0"
       assert query.select.params == []
-      assert query.select.take == %{0 => {:any, [:body, :title]}}
+      assert query.select.take == %{0 => {:any, [:dislikes, :title, :likes]}}
+
+      # On take with schemaless source
+      query = from c in "comments", select: [:title], select_merge: [:likes]
+
+      assert Macro.to_string(query.select.expr) == "&0"
+      assert query.select.params == []
+      assert query.select.take == %{0 => {:any, [:title, :likes]}}
+
+      # On take with schema
+      query = from c in Comment, select: [:title], select_merge: [:dislikes]
+
+      assert Macro.to_string(query.select.expr) == "&0"
+      assert query.select.params == []
+      assert query.select.take == %{0 => {:any, [:title, :dislikes]}}
     end
 
     test "on conflicting take" do