+++ test/ecto/query_test.exs
@@ -314,6 +314,12 @@ defmodule Ecto.QueryTest do
       assert %{post: 0} == query.aliases
       assert %{as: :post} = query.from
     end
+    
+    test "assigns a name using a variable" do
+      binding = :post
+      query = from p in "posts", as: ^binding
+      assert %{as: :post} = query.from
+    end
 
     test "assigns a name to a subquery source" do
       posts_query = from p in "posts"
@@ -325,7 +331,7 @@ defmodule Ecto.QueryTest do
 
     test "assign to source fails when non-atom name passed" do
       message = ~r/`as` must be a compile time atom or an interpolated value using \^, got: "post"/
-      assert_raise Ecto.Query.CompileError, message, fn -> 
+      assert_raise Ecto.Query.CompileError, message, fn ->
         quote_and_eval(from(p in "posts", as: "post"))
       end
     end
@@ -947,6 +953,73 @@ defmodule Ecto.QueryTest do
       assert_raise Protocol.UndefinedError,
                    ~r"protocol Ecto.Queryable not implemented for \[\]",
                    fn -> has_named_binding?([], :posts) end
+    end
+  end
+
+  describe "with_named_binding/3" do
+    test "executes a function with arity 1 when query does not have a named binding" do
+      query = from(p in "posts", as: :posts)
+
+      fun =
+        fn query ->
+          join(query, :left, [posts: posts], c in "comments", as: :comments)
+        end
+
+      query = with_named_binding(query, :comments, fun)
+
+      assert has_named_binding?(query, :comments)
+      assert %{joins: [%{as: :comments, source: {"comments", nil}}]} = query
+    end
+    
+    test "executes a function with arity 2 when query does not have a named binding" do
+      query = from(p in "posts", as: :posts)
+
+      fun =
+        fn query, binding ->
+          join(query, :left, [posts: posts], c in ^binding, as: ^binding)
+        end
+
+      query = with_named_binding(query, :comments, fun)
+
+      assert has_named_binding?(query, :comments)
+      assert %{joins: [%{as: :comments, source: {nil, :comments}}]} = query
+    end
+
+    test "does not execute a function when query has a named binding" do
+      query =
+        from(p in "posts", as: :posts,
+          join: c in "comments", as: :comments)
+
+      fun =
+        fn _query ->
+          raise "this function should not be called"
+        end
+
+      assert with_named_binding(query, :comments, fun) == query
+    end
+
+    test "raises when callback does not return Ecto.Query struct" do
+      assert_raise RuntimeError,
+                   ~r"callback function for with_named_binding/3 should return an Ecto.Query struct, got: :foo",
+                   fn -> with_named_binding(Schema, :comments, fn _query -> :foo end) end
+    end
+
+    test "raises when callback does not create a named binding" do
+      assert_raise RuntimeError,
+                   ~r"callback function for with_named_binding/3 should create a named binding for key :comments",
+                   fn -> with_named_binding(Schema, :comments, & &1) end
+    end
+    
+    test "raises when callback is not a function of arity 1 or 2" do
+      assert_raise ArgumentError,
+                   ~r"callback function for with_named_binding/3 should accept one or two arguments, got:",
+                   fn -> with_named_binding(Schema, :comments, &{&1, &2, &3}) end
+    end
+
+    test "casts queryable to query" do
+      assert_raise Protocol.UndefinedError,
+                   ~r"protocol Ecto.Queryable not implemented for \[\]",
+                   fn -> with_named_binding([], :posts, & &1) end
     end
   end